Try typecasting return to int from unsigned int

+++ 0x1E-search_algorithms/100-jump.c
@@ -30,7 +30,7 @@ int jump_search(int *array, size_t size, int value)
 	{
 		printf("Value checked array[%d] = [%d]\n", start, array[start]);
 		if (array[start] == value)
-			return (start);
+			return ((int)start);
 	}
 	return (-1);
 }
